requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
gunicorn>=20.0,<21.0
whitenoise>=5.2,<6.0
python-dotenv>=0.21,<0.22
Pillow